requirements.txt
asgiref==3.5.2
bleach==5.0.1
dj-database-url==1.0.0
Django==4.1
django-environ==0.9.0
django-markdownify==0.9.2
django-markdownx==4.0.0b1
Markdown==3.4.1
Pillow==9.2.0
Pygments==2.13.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
webencodings==0.5.1
whitenoise==6.2.0
